# api.yaml
openapi: 3.0.3
info:
  title: API for project 3d-inventory
  description: >-
    REST API for **3d-inventory**. Project is a simple solution that allows you to build a spatial and database
    representation of all types of warehouses and server rooms.
  contact:
    name: 3d-inventory-mongo-api
    url: https://github.com/karol-preiskorn/3d-inventory-mongo-api/
  version: 0.54.0
servers:
  - url: https://127.0.0.1:8080
  - url: http://0.0.0.0:8080
  - url: https://0.0.0.0:8080
  - url: http://localhost:8080
  - url: https://localhost:8080
  - url: http://127.0.0.1:8080
  - url: http://172.17.0.2:8080
  - url: http://172.17.0.3:8080
  - url: http://172.20.0.3:8080
  - url: https://d-inventory-api-wzwe3odv7q-ew.a.run.app
  - url: https://3d-inventory-api.ultimasolution.pl
  - url: https://3d-inventory.ultimasolution.pl
...
paths:
...
  /attributes:
    get:
      tags:
        - attributes
      summary: Get all attributes
      description: >-
        Retrieve a list of all attributes. You can apply optional query parameters such as `key` to filter
        attributes by their key or `type` to filter by their type.
      operationId: getAllAttributes
      security:
        - ViewerAuth: []
      responses:
        '200':
          description: A list of attributes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attributes'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (viewer role or higher required)
        '404':
          description: Not found.
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        '500':
          description: Unexpected server error
        default:
          description: Unexpected error
      x-swagger-router-controller: Attributes
    put:
      tags:
        - attributes
      summary: Update an existing attribute
      description: Update the details of an existing attribute.
      operationId: updateAttribute
      security:
        - UserAuth: []
      parameters:
        - name: id
          in: path
          description: ID of the attribute to update
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: 6566ed4ace63d42b25fbba19
            pattern: ^[0-9A-Fa-f]{24}$
            minLength: 24
            maxLength: 24
            x-42c-format: o:uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attribute'
        required: true
      responses:
        '200':
          description: Attribute updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (user role or higher required)
        '404':
          description: Attribute update data is inconsistent.
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        '500':
          description: Unexpected server error
        default:
          description: Unexpected error
      x-swagger-router-controller: Attributes
    post:
      tags:
        - attributes
      summary: Create a new attribute
      description: Create a new attribute. Provide attribute details in the request body.
      operationId: createAttribute
      security:
        - UserAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attribute'
      responses:
        '200':
          description: Attribute created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        '400':
          description: Bad request.
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (user role or higher required)
        '500':
          description: Unexpected server error
      x-swagger-router-controller: Attributes
    delete:
      tags:
        - attributes
      summary: Delete all attributes
      description: Remove all attributes from the inventory.
      operationId: deleteAllAttributes
      security:
        - AdminAuth: []
      responses:
        '200':
          description: All attributes deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmation'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: No attributes found to delete
      x-swagger-router-controller: Attributes
  /attributes/{id}:
    get:
      tags:
        - attributes
      summary: Get attribute by ID
      description: Retrieve a single attribute by its ID.
      operationId: getAttributeById
      parameters:
        - name: id
          in: path
          description: ID of attribute to fetch
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: 6566ed4ace63d42b25fbba19
            pattern: ^[0-9A-Fa-f]{24}$
            minLength: 24
            maxLength: 24
            x-42c-format: o:uuid
      responses:
        '200':
          description: Attribute found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        '404':
          description: Attribute not found
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        default:
          description: Unexpected error
      x-swagger-router-controller: Attributes
    put:
      tags:
        - attributes
      summary: Update attribute by ID
      description: Update a specific attribute by its ID.
      operationId: updateAttributeById
      parameters:
        - name: id
          in: path
          description: ID of attribute to update
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: 6566ed4ace63d42b25fbba19
            pattern: ^[0-9A-Fa-f]{24}$
            minLength: 24
            maxLength: 24
            x-42c-format: o:uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attribute'
      responses:
        '200':
          description: Attribute updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        '404':
          description: Attribute not found
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        default:
          description: Unexpected error
      x-swagger-router-controller: Attributes
    delete:
      tags:
        - attributes
      summary: Delete attribute by ID
      description: Delete a specific attribute by its ID.
      operationId: deleteAttributeById
      parameters:
        - name: id
          in: path
          description: ID of attribute to delete
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: 6566ed4ace63d42b25fbba19
            pattern: ^[0-9A-Fa-f]{24}$
            minLength: 24
            maxLength: 24
            x-42c-format: o:uuid
      responses:
        '200':
          description: Attribute deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmation'
        '404':
          description: Attribute not found
      x-swagger-router-controller: Attributes
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication. Include role and permissions for authorization.
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with admin role required
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with user role or higher required
    ViewerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with viewer role or higher required (read-only access)
  schemas:
...
    Attribute:
      description: >-
        Attribute is a key-value pair that can be associated with a device, model, or connection.
      x-42c-format: o:attribute
      required:
        - _id
        - attributeDictionaryId
        - value
      type: object
      properties:
        _id:
          example: 6566ed4ace63d42b25fbba19
          pattern: ^[0-9A-Fa-f]{24}$
          minLength: 24
          maxLength: 24
          x-42c-format: o:uuid
          type: string
        attributeDictionaryId:
          type: string
        connectionId:
          type: string
        deviceId:
          type: string
        modelId:
          type: string
        value:
          type: string
      example:
        _id: 653e10ecd6025afb66017c65
        attributeDictionaryId: 653e10ecd6025afb66017c65
        connectionId: 653e10ecd6025afb66017c70
        deviceId: 653e10ecd6025afb66017c71
        modelId: 653e10ecd6025afb66017c72
        value: 3W
    Attributes:
      type: array
      description: List of attributes
      maxItems: 1000
      example:
        - _id: 653e10ecd6025afb66017c65
          attributeDictionaryId: 653e10ecd6025afb66017c65
          connectionId: 653e10ecd6025afb66017c70
          deviceId: 653e10ecd6025afb66017c71
          modelId: 653e10ecd6025afb66017c72
          name: power
          value: 3W
        - _id: 653e10ecd6025afb66017c66
          attributeDictionaryId: 653e10ecd6025afb66017c66
          connectionId: 653e10ecd6025afb66017c73
          deviceId: 653e10ecd6025afb66017c74
          modelId: 653e10ecd6025afb66017c75
          name: color
          value: black
        - _id: 653e10ecd6025afb66017c67
          attributeDictionaryId: 653e10ecd6025afb66017c67
          connectionId: 653e10ecd6025afb66017c76
          deviceId: 653e10ecd6025afb66017c77
          modelId: 653e10ecd6025afb66017c78
          name: year
          value: '1999'
      items:
        $ref: '#/components/schemas/Attribute'
...
    Confirmation:
      type: object
      properties:
        message:
          type: string
      additionalProperties: false
      description: Universal confirmation message from database
      example:
        message: Succeed
...